add project delete validation

+++ src/components/ProjectDelete.js
@@ -8,7 +8,7 @@ import "../styles/deleteProject.css"
 export const ProjectDelete = (props) => {
   const [text, setText] = useState("")
   // const [accountForPosting, setAccountForPosting] = useState(props.settings.strings["only_users_can_post"])
-
+  const [validation, setValidation] = useState(false);
   const [project, setProject] = useState(props.projects.active)
 
   //in useEffect, check if we have some active project
@@ -21,22 +21,18 @@ export const ProjectDelete = (props) => {
 
   const deleteProjectClick = (event) => {
     event.preventDefault()
+    console.log(props)
+    if (text === props.settings.strings["confirm_delete_text"]) {
+      setValidation(false)
+      let new_mods = []
+      new_mods.push(props.user.username)
 
-    let new_mods = []
-    new_mods.push(props.user.username)
-    const object = {
-      "id": 123,
-      "lang": "en",
-      "moderators": new_mods,
-      "info": {
-        "lang": "en",
-      },
-      "admins": new_mods
+      props.deleteProject(project.id)
+      console.log(project)
+      props.notify(`Project "${project.title}" Delete.`, 5)
+      // props.history.push("/")
     }
-    props.deleteProject(project.id)
-    console.log(props, project)
-    // props.notify(`Project "${title.title}" Delete.`, 5)
-    // props.history.push("/")
+    else setValidation(true)
   }
 
   const textChangeHandler = (event) => {
@@ -52,7 +48,9 @@ export const ProjectDelete = (props) => {
           <div className="inputContainer">
             <input name="text" className="input" placeholder={props.settings.strings["confirm_delete_text"]} maxLength="32" onChange={textChangeHandler} value={text} />
             <div className="inputFocusLine" />
+            {validation && <div className="error">"Wrong input, write again"</div>}
           </div>
+
         </div>
 
       </form>
@@ -87,5 +85,6 @@ export default connect(
 
 
 
-// DeleteProject: The DeleteProject component is defined, which renders a form that ask user to type 'comfirm delete', preventing user from unwanted deletion.
+// DeleteProject: The DeleteProject component is defined, which renders a form that ask user to type 'comfirm delete', preventing authorized user from unwanted deletion.
 // The state of the component is managed using the useState hook.
+// The current delete is appeard to all project's admin and moderators. Need to check whether it is user owner or not (only owner can delete).
